+++ extensions/launchdarkly/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Launchdarkly Changelog
 
+## [Small fixes] - 2025-02-08
+
+- 👥 Avatar icons are now displayed better using the `getAvatarIcon` function from `@raycast/utils`
+- 📋 Copy feature flag key to clipboard
+- 🔗 Expand open in browser action in the feature flag details
+- 📝 Updated readme
+
 ## [Initial Version] - 2025-01-10
 
 - 🔍 Search through all your feature flags